+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- fix: Transaction default status should be OK (#1439)
+
 ## 7.5.0
 
 - feat: Add one flag to disable all swizzling (#1430)